fix: update price calculation to use floats and bump version to 1.8.4
- Changed price_in_tomans calculation in purchase_plan.py to use float instead of int for accurate handling of exchange rates and prices, preventing potential truncation errors.
- Updated VERSION from 1.8.3 to 1.8.4.

+++ core/scripts/telegrambot/utils/purchase_plan.py
@@ -308,7 +308,7 @@ def handle_card_to_card_payment(call, plan_gb):
         price = plan['price']
         
         # Convert price to tomans using the exchange rate
-        price_in_tomans = int(price) * int(exchange_rate)
+        price_in_tomans = float(price) * float(exchange_rate)
 
         message = get_message_text(language, "card_to_card_payment").format(price=price_in_tomans, card_number=card_number)
